CrossWordSolvingAlgos
These algorithms implement several basic search algorithms based on a tree structure. In this case they are used to solve a 3*3 crossword. DepthFirstSearch and BreadthFirstSearch are both are fairly self explanatory. DepthFirst starts at the root of the tree and recursively finds the children of the root. It then immediately finds the children of the first child until the offspring have no children. Offspring with no children are checked before more of the tree is explored. Breadth first instead finds all parent nodes, moving across an entire level of the tree before recursing down a level. The primary difference between these two implementations is that DepthFirst uses a stack to store nodes while while BreadthFirst uses a queue. The HeuristicSearch is modification of DepthFirst that sorts the "grandchildren" of the root by the number of vowels that a particular state of the crossword possesses and checks states with more vowels first.  The vowel heuristic may be useful in some cases but its debatable.